Reverting resources app search indices to its own file. This commit undoes 855b798a435c4349e3cf494963e81a3d6b936c94, which seems to have broken solr indexing for resources.
